refactor(data): extract registry/detail sources
